Removed some unnecessary code bits
All of these values were already set in the base weapon and were exactly the same in every single weapon, so they should just make use of the base's value instead.

Also moved the ACF logo material to a slightly better location and fixed a bug that allowed you to fire while sprinting with the shotgun.